# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Portfolio Tracker", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Root", "description": "Root endpoint - basic info.", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/health": {"get": {"tags": ["window"], "summary": "Window Health Check", "description": "Check if the Tauri window socket server is available.", "operationId": "window_health_check_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Window Health Check Api Health Get"}}}}}}}, "/api/assets/": {"get": {"tags": ["assets"], "summary": "List Assets", "operationId": "list_assets_api_assets__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response List Assets Api Assets  Get"}}}}}}, "post": {"tags": ["assets"], "summary": "Create Asset", "operationId": "create_asset_api_assets__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AssetCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Create Asset Api Assets  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/assets/{asset_id}": {"get": {"tags": ["assets"], "summary": "Get Asset", "operationId": "get_asset_api_assets__asset_id__get", "parameters": [{"name": "asset_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Asset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Get Asset Api Assets  Asset Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["assets"], "summary": "Update Asset", "operationId": "update_asset_api_assets__asset_id__put", "parameters": [{"name": "asset_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Asset Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AssetUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Update Asset Api Assets  Asset Id  Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/exchanges/": {"get": {"tags": ["exchanges"], "summary": "List Exchanges", "operationId": "list_exchanges_api_exchanges__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response List Exchanges Api Exchanges  Get"}}}}}}, "post": {"tags": ["exchanges"], "summary": "Create Exchange", "operationId": "create_exchange_api_exchanges__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExchangeCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Create Exchange Api Exchanges  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/exchanges/{exchange_id}": {"delete": {"tags": ["exchanges"], "summary": "Delete Exchange", "operationId": "delete_exchange_api_exchanges__exchange_id__delete", "parameters": [{"name": "exchange_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Exchange Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Delete Exchange Api Exchanges  Exchange Id  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/transactions/": {"post": {"tags": ["transactions"], "summary": "Create Transaction", "operationId": "create_transaction_api_transactions__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Create Transaction Api Transactions  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/transactions/{asset_id}": {"get": {"tags": ["transactions"], "summary": "List Transactions", "operationId": "list_transactions_api_transactions__asset_id__get", "parameters": [{"name": "asset_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Asset Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response List Transactions Api Transactions  Asset Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/transactions/{tx_id}/update": {"put": {"tags": ["transactions"], "summary": "Update Transaction", "operationId": "update_transaction_api_transactions__tx_id__update_put", "parameters": [{"name": "tx_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Tx Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Update Transaction Api Transactions  Tx Id  Update Put"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/transactions/{tx_id}/delete": {"delete": {"tags": ["transactions"], "summary": "Delete Transaction", "operationId": "delete_transaction_api_transactions__tx_id__delete_delete", "parameters": [{"name": "tx_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Tx Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Delete Transaction Api Transactions  Tx Id  Delete Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/positions/": {"get": {"tags": ["positions"], "summary": "Get Positions", "operationId": "get_positions_api_positions__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Get Positions Api Positions  Get"}}}}}}}, "/api/prices/refresh": {"post": {"tags": ["prices"], "summary": "Refresh Prices", "operationId": "refresh_prices_api_prices_refresh_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object", "title": "Response Refresh Prices Api Prices Refresh Post"}}}}}}}, "/api/net-worth/": {"get": {"tags": ["net-worth"], "summary": "Get Net Worth", "operationId": "get_net_worth_api_net_worth__get", "parameters": [{"name": "period", "in": "query", "required": false, "schema": {"type": "string", "default": "1m", "title": "Period"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Get Net Worth Api Net Worth  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/export/": {"get": {"tags": ["export"], "summary": "Export Db", "operationId": "export_db_api_export__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/secrets/{name}": {"post": {"tags": ["secrets"], "summary": "Set Secret", "operationId": "set_secret_api_secrets__name__post", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SecretSet"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Set Secret Api Secrets  Name  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["secrets"], "summary": "Delete Secret", "operationId": "delete_secret_api_secrets__name__delete", "parameters": [{"name": "name", "in": "path", "required": true, "schema": {"type": "string", "title": "Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Response Delete Secret Api Secrets  Name  Delete"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/window": {"post": {"tags": ["window"], "summary": "Toggle Window State", "description": "Toggle window maximize/restore state via Unix socket to Tauri.\n\nThis endpoint communicates with the Rust backend through a Unix socket\nto control the window state (maximize/restore).", "operationId": "toggle_window_state_api_window_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WindowStateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Toggle Window State Api Window Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AssetCreate": {"properties": {"symbol": {"type": "string", "maxLength": 50, "minLength": 1, "title": "Symbol"}, "name": {"type": "string", "maxLength": 255, "minLength": 1, "title": "Name"}, "type": {"type": "string", "title": "Type"}, "exchange_id": {"type": "integer", "title": "Exchange Id"}, "yahoo_ticker": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Yahoo Ticker"}, "coingecko_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Coingecko Id"}}, "type": "object", "required": ["symbol", "name", "type", "exchange_id"], "title": "AssetCreate"}, "AssetUpdate": {"properties": {"symbol": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Symbol"}, "name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Type"}, "exchange_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Exchange Id"}, "yahoo_ticker": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Yahoo Ticker"}, "coingecko_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Coingecko Id"}}, "type": "object", "title": "AssetUpdate"}, "ExchangeCreate": {"properties": {"name": {"type": "string", "maxLength": 255, "minLength": 1, "title": "Name"}, "type": {"type": "string", "title": "Type"}}, "type": "object", "required": ["name", "type"], "title": "ExchangeCreate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "SecretSet": {"properties": {"value": {"type": "string", "title": "Value"}}, "type": "object", "required": ["value"], "title": "SecretSet"}, "TransactionCreate": {"properties": {"asset_id": {"type": "integer", "title": "Asset Id"}, "date": {"type": "string", "title": "Date"}, "type": {"type": "string", "title": "Type"}, "units": {"type": "number", "title": "Units"}, "eur_amount": {"type": "number", "title": "Eur Amount"}, "notes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Notes"}}, "type": "object", "required": ["asset_id", "date", "type", "units", "eur_amount"], "title": "TransactionCreate"}, "TransactionUpdate": {"properties": {"date": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Date"}, "type": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Type"}, "units": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Units"}, "eur_amount": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Eur Amount"}, "notes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Notes"}}, "type": "object", "title": "TransactionUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WindowStateRequest": {"properties": {"action": {"type": "string", "title": "Action"}}, "type": "object", "required": ["action"], "title": "WindowStateRequest", "description": "Request model for window state control."}}}}